moved components to routes
